requirements.txt
kafka-python-ng
py-jsonl
